requirements.txt
fastapi
uvicorn[standard]
python-multipart
spacy
scikit-learn
sentence-transformers
pdfplumber
python-docx
sqlalchemy
